Fixed notification action for terminated app. (#50)
Added "FirebaseMessaging.instance.getInitialMessage" to handle the action for a message/notification that was used to open the terminated app.

When opening a notification (tapping on it) we can specify an action that should happen (e.g. navigating to a page or website) (see #3).

This worked for an app running in the foreground or background but not an app that was terminated.

What was missing was a check if "FirebaseMessaging.instance.getInitialMessage" returns a message and if so handling this message.
This behavior was added with this commit.