Recreate Podfile to start fix macOS build. (#64)
Building macOS on `main` resulted in the following error
```
[!] No podspec found for `path_provider` in `Flutter/ephemeral/.symlinks/plugins/path_provider/macos`
```

This seems to come from an old macOS Podfile. When Flutter creates a new Podfile many lines of code are not existing that we had in the Podfile in main. 

With this PR I:
* Created a new Podfile by deleting the old one and letting Flutter create a new version (by running `flutter pub get`).
* Add the build time fix for Firestore.
* Update Firestore dependency (old version threw an error when running `pod install`)

With these fixes I can now build the `macOS` version again on my MacBook.

Co-authored-by: nilsreichardt <nils@sharezone.net>
Co-authored-by: nilsreichardt <nils@reichardt.io>