Add banner when app is running an alpha version (#216)
## Description
This PR contains a proposal to display an alpha version banner.

Users which are using an alpha version should be aware that is an alpha version. Therefore, a small banner at the top right-hand corner indicates this.

This banner also helps us when somebody sends a screenshot of an alpha version. We can instantly see that is an alpha version.

When this proposal got accepted, we can add the environment variable to the `deploy_web_app` workflow.

## Demo
<img width="1135" alt="image" src="https://user-images.githubusercontent.com/24459435/167873877-2a0f512b-14c9-4ee9-a0f8-5662c5c26b52.png">

## Related Tickets
Related to #27

The PR #220 should be merged first.