Add automatically labels to pull requests based on the changed file paths (#229)
## Description
Finding pull requests with GitHub can be hard. Especially when our app has a bug and we try to find the pull request which caused the bug, might it be helpful to labels, like `platform: android` when we know it something with the Android build or changes to our CI/CD with the label `ci/cd`.

Therefore, I added the `labeler` Action to automatically add labels to our pull requests based on the changed file paths.