Add Codemagic App Previews (Android & iOS) for Pull Requests (#257)
## Description

We are already using the [Deploy to
Firebase](https://github.com/marketplace/actions/deploy-to-firebase-hosting)
action to deploy our web to a temporary url, when opening a new pull
request.

[codemagic_app_preview](https://pub.dev/packages/codemagic_app_preview)
does the same for iOS and Android. Just scan the QR with your phone,
install the app and review the changes 👍

Currently, I'm not able to build a iOS and macOS with one Codemagic
workflow. Therefore, we are not building the macOS app at the moment.
When I fixed this, I will make a new PR for adding support for macOS 👍

Google Sign In & Dynamic Links are not working at the moment for the
Android version, because the dev Firebase project does not contain the
SHA1 & SHA256 values of the code signing key (see #289).

Co-authored-by: Jonas Sander <29028262+Jonas-Sander@users.noreply.github.com>