Scroll automatically section right up to threshold
While developing I used a threshold that was near the top of the
page but not at the very top `threshold: 0.1`. Automatic
scrolling (i.e. tapping on a section heading in the table of
contents) would scroll the heading to the very top though back
then.

This lead to the case that automatically scrolling to a section
heading of a very small heading might skip this section as the
heading of the next section was already in the threshold.

With this test we ensure that we scroll the heading always right
up to the threshold.