Refactor group permission function.  (#332)
I stumbled on the `requestPermission` function in our code which was
used to check if a user can perform a certain action, e.g. adding
homework or changing course settings.

This function was weirdly named and had logic for enum values which were
not in use anymore.

I changed the function to an extension method on `MemberRole` and
removed the unused enums.
Additionally I moved the files from the `additional` to `groups` folder.
Lastly I renamed the function `isUserAdminOrOwnerFromCourse` to
`isUserAdminOrOwnerOfGroup` in the same file.