Refactor holiday code into separate holidays package (#324)
Moved most of the holiday code to `lib/holidays`.  
I left holidays_bloc inside of the `app` code since it interacts with
the `UserGateway` and is more specific to the dashboard page widget.
I didn't change any logic.  
I ran the app manually on the mac to check if it still works.