Use `fvm` in GitHub Actions and Sharezone CLI. (#331)
Instead of using `flutter` and `dart` we use `fvm flutter` and `fvm
dart` in the CLI and GitHub Actions to ensure the same version of
Flutter and Dart is used everywhere.

As a quick fix I just replaced the commands, if `fvm` is not installed
this will fail instead of falling back to using just `flutter`/`dart`.
We might add this fallback mechanism later if we really need it.

I tested this change via `sz test` and `sz analyze` locally.

Fixes #255