Fix macOS build after issue with `pdfx` (#354)
## Description

This PR fixes the build issue for macOS (#353). The main problem is an
issue in the `pdfx` package:
https://github.com/ScerIO/packages.flutter/issues/358. To fix this
issue, we are using a PR with fixes the issue:
https://github.com/ScerIO/packages.flutter/pull/384.

Additionally, I needed to upgrade `device_info_plus` to the newest
version because `pdfx` depended on a newer version. Furthermore, I need
to upgrade the Gradle version to 7.3.3 (required by `device_info_plus`).

## Related Tickets

Closes #353