Add `sharezone_lints` package (#358)
## Description

Adds the `sharezone_lints` package to our `lib` folder. This package
contains just a `analysis_options.yaml`, which redirects to the
`flutter_lints` package.

The advantage of this package is that we can this package to our
internal package (in future PRs) to have the same lint rules across the
whole repository. Additionally, adding a new lint rule would also
enforce to apply this lint to all package to have a consistent code
style.

As already mentioned, we are going to add the `sharezone_lints` package
in future PRs. This is because these PRs will produce a lot of line
changes and splitting this into multiple PRs makes it much easier to
review.

Closes #316