Initialize Firebase via Dart (#376)
## Description

With this PR we initialize Firebase also via Dart. This change requires
setting the `main` file when running the Flutter app:

```sh
flutter run \
  --flavor dev \
  --target lib/main_dev.dart
```

This is because we need a way to know which flavor is used inside the
Dart code to terminate which Firebase project we should initialize. An
alternative would be to use `--dart-define`. However, I like the
`--target` solution more because there is no default option and you have
to explicitly set a flavor. With `--dart-define`, we would either need
to set a default or throw an exception at runtime.

Our file structure looks like this:
```
lib/
  ...
  main_common.dart
  main_dev.dart
  main_prod.dart
```

`main_common.dart` is just the old `main.dart` but with a required
`flavor` parameter.

We are going to keep the native Firebase initialization for Android on
iOS because we need it for Crashlytics and from my experience is the
initialize Firebase way for Android and iOS not that stable. It's no
problem to have both solutions.

## Related Tickets

Closes #375 
Closes #221