Split `main.yml` into `app_ci.yml` and `license.yml` to use `path` filters (#400)
Before this PR we had all CI workflow in one `main.yml` file. The
disadvantage: we always triggered the unit tests, analyze jobs,
integration tests, web preview, etc. even when only made a change in
`.md` file -> always long CI times.

To be able to only run the app workflow for app changes, we need to
define the triggers on a file level. Therefore, we needed to split the
workflows in separated file because sometimes we want to run workflow
always (like the license check workflow) and some action only when the
app changed (the `analyze` job).

We have now these file structures:
* `app_ci.yml`: `analyze`, `test`, `web-preview`,
`ios-integration-test`, `android-integration-test`
* `license.yml`: `check-files-licence-headers`