Add Fallback App CI (#404)
![image](https://user-images.githubusercontent.com/24459435/221893313-7aa00dea-a22d-4280-a05f-242586596c12.png)


A workaround for GitHub Actions jobs that required and based on a path.

Currently, it's not possible to set a GitHub Action to required which is
based on path changes. The reason for this that the action is not
executed when the action does not match the paths. Therefore, merging
will be blocked.

A workaround is to have a job which the same name. 

See:
https://github.com/github/docs/commit/4364076e0fb56c2579ae90cd048939eaa2c18954

Closes #403