Remove pull request trigger `codemagic.yaml` (#410)
Removed the pull request trigger for `codemagic.yaml` because it's not
worth.

I kept the pull request trigger for the file `codemagic.yaml` in the
`codemagic.yaml` to ensure that the `codemagic.yaml` works when you
change something. However, I think it's a bit overkill because Codemagic
needs to spin up a macOS VM and clone the repository to be able to check
if the `codemagic.yaml` has changed (and this for *every* commit). This
process eats ups our Codemagic minutes.

We can test the `app-preview` workflow in the `codemagic.yaml` by adding
the `build-app-preview` label 👍