Use `dorny/paths-filter` instead of native path filters for our CI (#456)
Because `merge_groups` does not support paths, we try to use the
`dorny/paths-filter` action (workaround from
https://github.com/community/community/discussions/45899).

We define in the `changes` check when `changeFound` should be true. All
other checks should wait until `path-filter` is finished (`needs:
changes`). When other checks are executed, they need to check if
`changeFound` is true. If yes, execute the check, if no, skip.


![image](https://user-images.githubusercontent.com/24459435/224088630-41274668-7b9a-4eea-ba86-e2adcb940047.png)

With this workaround for the `merge_groups`, can we also remove our
workaround for required checks that executed based on a path (fallback
CI are the recommend workaround from
https://github.com/github/docs/commit/4364076e0fb56c2579ae90cd048939eaa2c18954)