Bump build number for Android alpha builds (#471)
## Description

On Android it's hard to say in which version exactly a bug occurred (or
in version the app is working again).

Therefore, we are bumping now the Android build number. We use for this
the latest AppStore build number (as SOT for the build number) because
there is no easy API for Firebase Distribution to get the latest build
number. The only cases where this could be a problem would be when the
iOS build fails (this would result in a duplicated build number on
Android). However, even when this happens is that not dramatic because
the build number is just for us and is not required to be increased for
every build.

Problem on Android (it's always 1.6.2 (318)):


https://user-images.githubusercontent.com/24459435/224200019-99127317-bcd4-4bb9-9e49-42c1130e2901.mp4