Extend `labeler.yml` to automatically assign nearly all labels (#493)
This PR adds all of our feature labels and a few more labels to our
`labeler.yml` with the goal that 95% of the labeling is done
automatically.

If we don't automate this process, our PR will have a bad quality of
labels.

We try to use general and wide globs for assigning labels. For the
"feature: homework" label, do not list all the paths to the homework
directories. Instead, just use something like `**/homework/**`. This
will automatically include the app files, test files, package files and
asset files that contain a homework directory. Of course, this can lead
to false positives in a few cases. However, we prefer the general glob
to the small chance of a false positive because we think the chance of
not assigning the label with too specific globs is much higher than
having a false positive with a general glob. And when searching for a PR
with labels, it is usually better to find a few false positives than not
to find a PR about a change in the homework feature that does not have
the homework label.

We are not going to include "w: " labels because they are too specific
and it's too hard to maintain the paths for the files. Searching stuff
with the "feature: " label should be enough.