Remove deprecated Android `SplashScreenDrawable`.
We still used an deprecated way to declare a splash screen, which caused the following warning to be printed on Android start-up: `W/FlutterActivityAndFragmentDelegate( 6179): A splash screen was provided to Flutter, but this is deprecated. See flutter.dev/go/android-splash-migration for migration steps.`

Since we also already use the recommand way to declare a splash screen additionally, we just had to remove the old way. There are no behavior / splash screen changes, everything works as it used to.