Cancel CI for outdated commits in a PR (#560)
Because we had set the concurrency for the merge queue to 1 (because
setting the value to > 1 has no effect because we only have 5 macOS
machines), we are never running two 2 PRs at the same time. The reason
why we removed this feature was that when 2 PRs were actively in the
merge queue (running, not waiting) at the same time, the PRs were
cancelled. So I think we add again the cancel-in-progress feature (I'm
not 100% sure if it's going to work but I think it should).

<img width="430" alt="image"
src="https://user-images.githubusercontent.com/24459435/229933740-9b0c75d7-f0f9-4ec1-a83a-8bf8ae088697.png">