Fix the file upload for the web app (#440)
## Description

This PR fixes the file upload for web app. The issue was #369 removed
the file uploader for web because we removed the dependency of the
`firebase` package.

Since the official `firebase_storage` is available for web, we can use
the method from the official package. However, for web can't use
`putFile` because it's not available for web. In this case, we need to
use `putData`.

## Demo

Filesharing:


https://user-images.githubusercontent.com/24459435/223853461-2b24ffb9-f2e4-439d-b526-d9cfaf8d8506.mov

Homework submissions



https://user-images.githubusercontent.com/24459435/223853933-f81ab5c7-0ceb-4f5a-b59d-b1e279cff724.mov

(the cold start, lol)

## Related Tickets

Closes #374

---------

Co-authored-by: Jonas Sander <29028262+Jonas-Sander@users.noreply.github.com>