Add "Sharezone Plus" subscription prototype (#552)
This adds the first prototype of a "Sharezone Plus" subscription.

The prototype can be enabled by typing "SharezonePlus" into the textbox
at "Einstellungen" > "Mein Konto" > "Aktivierungscode eingeben".

If the prototype is enabled and a teacher hasn't made a subscription
then the accessing the submissions via the homework details page will be
disabled (one will be redirected to the subscription page). If the
subscription has been made then the user can access the submissions page
as usual.

This prototype will probably only work on Android.


https://user-images.githubusercontent.com/29028262/230096282-3ac40125-933f-4cf6-af93-2520a1dcadb3.mp4

---------

Co-authored-by: nilsreichardt <me@nils.re>
Co-authored-by: Jonas Sander <Jonas-Sander@users.noreply.github.com>