Split our CI test job into `test-without-goldens` and `test-only-goldens`  (#617)
The only reason why used a macOS runner to run our tests was that we
have few golden tests which require a macOS runner. Since we have
`--exclude-goldens` and `--only-goldens` arguments for the `sz test`
command (see #614), we can split the test jobs into a job that runs only
the golden tests (uses a macOS runner) and a job that excludes the
golden tests (uses an Ubuntu runner).

This PR also uses the `--exclude-goldens` argument for the CLI CI
because the CLI has no golden tests.