Fix loading homeworks for teachers (#655)
## Summary

This PR fixes:

* #608 
* #654 
* #497 (real fix)

It's required to this all in one pull request because otherwise it would
introduce a new bug.

## What was the issue?

I started to debug on this problem because of #654. I found out it was
because `TypeOfUser.student` was used as a backup when it's null (see
#573). However, just reverting it to `TypeOfUser.unknown` was not option
because this would introduce #497 again.

Therefore, I wondered, why `typeOfUser` can be null. The reason for this
was the last time online reporter. Because `.set()` has been used for
Firestore, Firestore instantly was setting it null (I think `null` is
used for `Firestore.currentTimestamp()` when the client is offline or
the server hasn't responded yet). Using `.update()` fixed the issue that
the user document only contained `lastOnline`.

But using `update()` in the last online reporter always caused #497.
Even using `TypeOfUser.student` or `TypeOfUser.unknown` couldn't be as a
workaround. After hours of debugging I found that when signing our the
following line was causing the issue:


https://github.com/SharezoneApp/sharezone-app/blob/47e4ccb0993f01b5c9b8c83f31f7ec4c4f833c77/app/lib/util/api/homework_api.dart#L107

The issue is that we listened to `_homeworkSubjectStream` and when an
error occurred, we the `.addError` method of `_homeworkSubjectStream `
used. This ended up into an infinity loop which resulted in the freeze
(#497),

This was by far one of my hardest bugs... I needed ~5,5h to find the
solution and it was extremely hard to debug.

Fixes #608
Fixes #654
Fixes #497