Add CD workflow to deploy new iOS stable versions (#699)
This pull request introduces a new GitHub Actions workflow for deploying
new stable versions of our application.

The workflow, named 'stable', ensures that our iOS builds are built
sequentially to maintain unique build numbers, a requirement of iOS
builds. The concurrency group, 'app-release', enforces that only one
version is built per branch at a time.

A manual trigger has been implemented via the workflow_dispatch event.
This requires the input of a changelog for the iOS release. This
changelog is subsequently used as the release notes in the App Store
release.