Skip image compressing for macOS (#714)
## Summary

This PR addresses a limitation with the `flutter_image_compress` library
on macOS. Due to a known issue, the image compressor does not work on
macOS, leading to potential runtime errors or unexpected behavior.

## Changes:

- If the platform is macOS, the image compression step is skipped and
the original file is returned unchanged.

## Reason for Change:

The `flutter_image_compress` library has a known issue on macOS which
causes the compression functionality to fail. Refer to the following
issue for more details:
[fluttercandies/flutter_image_compress#255](https://github.com/fluttercandies/flutter_image_compress/issues/255).

By skipping the image compression step on macOS, we ensure that our
users do not encounter unexpected errors or issues while using the app
on this platform. The changes are a temporary measure until a fix or
alternative solution is found for the underlying library issue.

## Testing:

- Manual testing on macOS confirms that images are not compressed and
the app behaves as expected.

## Related Tickets

Fixes #713