Do not store an event when start and end time is equal or end is before start time (#777)
## Description

Prior to this PR, when adding an event where the start and end times are
the same or the end time is before the start time, the application would
simply close the event creation dialogue and display a confirmation
message. This PR fixes this behavior and also adds a snack bar with a
helpful message if the wrong values are selected.

Code quality is not the best, just sticking to the current approach
because otherwise a big refactoring would be needed.

## Demo


https://github.com/SharezoneApp/sharezone-app/assets/24459435/af91eec5-bbd2-459b-9cd7-1e89659f7701

## Related tickets

Fixes #107