Fix opening links on Android (#791)
As in #790 described, opening links on Android doesn't work anymore. I
tried it with two phones: Pixel 6a and Pixel 4a. When I removed the
`canLaunch` check, it worked again. Additionally, I added
`LaunchMode.externalApplication` to prevent opening links in an in-app
web view. This is required to open the apps of the link, like the Play
Store in case for #782.

Fixes #790