Move golden tests of `qr_code_scanner` into the `test_goldens` folder (#794)
I forget to change to move the golden test into `test_goldens` so that
`sz test --only-goldens` also tests this package. But it wasn't that
dramatic because the generated images looked exactly on Linux, which
what the reason that the CI pipeline not failed. Just making this change
to be consistent.