Show snack bar when changing course color was not successful (#799)
When changing the color, the server could return `false` when for some
reason the change was not successful (ideally, the server would send /
throw an exception with more information). However, we when the change
was not successful, we have just shown "Changing was successful" what is
not true. With this PR we show that the change was not successful. In a
different PR I'm going to fix #798 (this is PR is only related to #798)
so that we show something like a lock icon when the user has no
permission.


![image](https://github.com/SharezoneApp/sharezone-app/assets/24459435/ce919422-2fbd-4431-be75-adfe969dc9d2)

Related to #798