Use `TextCapitalization` (#829)
Adds `TextCapitalization` to text field where
`TextCapitalization.sentences` or `TextCapitalization.characters` could
be helpful. `TextCapitalization.sentences` starts the first character
with an upper case characters. This can be helpful for the text field
where I added it because in most cases you want to start a with an upper
case character. `TextCapitalization.characters` writes every character
as upper case character. It can be helpful for abbreviations.

This change only effects Android and iOS.