Make Flutter to non-fatal issues (#857)
We marked our Flutter exceptions as fatal issues, which led to a high
crash rate in Crashlytics, because Crashlytics thought that all
exceptions were crashes. This is not good because now we can't determine
which ones are really crashes. However, we need to investigate here
because normally Flutter errors should be caught via
`FlutterError.onError` 🤔