Implement Sharezone Plus page (#898)
## Description

This PR adds the implements the design of the Sharezone Plus page. I
also added a bit of logic for the page (`SharezonePlusPageBloc)`.

The following things will be implemented later: 
* `buySubscription` and `cancelSubscription` methods in the
`SharezonePageBloc`.
* Analytics for the page

I only added the advantage of open source. The other advantages will be
added when this plus feature is implemented.

## Demo

### Light Mode


https://github.com/SharezoneApp/sharezone-app/assets/24459435/55926759-d03b-4a09-b9e9-257ae9329059

### Dark Mode


https://github.com/SharezoneApp/sharezone-app/assets/24459435/a0a43e11-45dd-4e14-8e51-6462e6de2827

## Testing

I added golden tests for light and dark mode and every screen layout.

## Related Tickets

Closes #893