Add signing steps to the `sz deploy ios` command (#911)
Currently, we run the signing steps in the GitHub Actions file. However,
we can't do this for the macOS release because signing is part of the
deploy process. As prework, I added the signing steps to the iOS command
because the signing for iOS is easier than for macOS. When this works,
can we make a new PR for the `sz deploy macos` command.

Manually tested it with:
https://github.com/SharezoneApp/sharezone-app/actions/runs/6119211836/job/16608705828

---------

Co-authored-by: Jonas Sander <29028262+Jonas-Sander@users.noreply.github.com>