Restrict access for filter by school class for Sharezone Plus users (#945)
## Description

This PR restricts access to the school class filters for Sharezone Plus
users. At the moment it just redirects to the Sharezone Plus page if the
user doesn't have Sharezone Plus. But I want to show a dialog with a
video demo first, so that the user can verify "Ah, yes, that's the
feature for what I'm looking for". Will be implemented in a future PR
(#953).

## Demo

Desktop:


https://github.com/SharezoneApp/sharezone-app/assets/24459435/42c76e9f-79eb-482f-a1f7-65f8ec09a73a

Mobile:


![image](https://github.com/SharezoneApp/sharezone-app/assets/24459435/41f4a0a9-abfe-4395-8f30-cf4e48799a2b)

## Related Tickets

Closes #852