Fix Multiline Release Notes Handling in GitHub Action (#980)
In our previous GitHub Action configuration, the `--whats-new` parameter
was passed with an equals symbol (`=`), as in
`--whats-new="$SHORT_LAST_COMMIT_MESSAGE"`. However, this approach
encountered issues when the release notes contained a multiline string,
causing the multiline string not to be passed correctly to the Dart CLI
tool, resulting in a null value received in the Dart script.

To resolve this, we modified the syntax to `--whats-new
"$SHORT_LAST_COMMIT_MESSAGE"`, omitting the `=`. This ensures that
multiline strings are correctly passed as a single argument to the CLI
tool, preserving the integrity of the release notes.

This update ensures that release notes for alpha versions generated from
multiline commit messages are correctly handled and displayed.

Also reverts the changes from #975

Fixes #922