Add Sharezone Plus support (#962)
## Description

This adds the plus support to our support page.

### Challenges

When implementing this feature, I had the following challenge: We also
use the support page in areas of the app, where the user isn't signed in
and can't buy Sharezone Plus and `SubscriptionService` and
`SubscriptionEnabledFlag` aren't available.

I solved it by:

* Initializing the `SupportPageController` at two different locations:
`SharezoneApp` (user is authenticated) and `AuthApp` (user is not
authenticated)
* Therefore, I can set the attributes of the `SupportPageController` to
`Stream.value(false)`
* I moved `SubscriptionEnabledFlag` above `SharezoneApp` and `AuthApp`
to be able to access it in both subtrees.

## Demo


![image](https://github.com/SharezoneApp/sharezone-app/assets/24459435/f7a55c80-1b46-4805-8b88-db8f78a603b7)

## Related Tickets

Closes #885