Going back to `macos-12` for iOS integration tests to improve the performance dramatically  (#1012)
I found the issue why our build times for iOS integration tests went
from 15 - 20 minutes to 60 - 100 minutes: it's `macos-13`. Going back to
`macos-12` solves this issue. Invertase faced the same issue:
https://github.com/firebase/flutterfire/pull/11451.

I tested it with 3 builds using `macos-12`. All three only took 23
minutes:


![image](https://github.com/SharezoneApp/sharezone-app/assets/24459435/9579f51b-1b19-4784-aaf3-98f7e7bb4d56)


https://github.com/SharezoneApp/sharezone-app/actions/runs/6228178121/job/16904381314?pr=1012

However, using `macos-12` has the problem that `macos-12` does not
support Metal (https://github.com/flutter/flutter/issues/126768), which
is used by Impeller. So we have to disable Impeller for integration
testing. This is not really great, because in production we use
Impeller, and there could be a problem with Impeller in our application
that is not tested. In the end, it's a trade-off. I think the
probability that there is an issue with Impeller that is covered by our
integration tests is very low, and having 120 minutes of integration
tests really slows down development. So I would choose the option to
disable Impeller for now.