Add option to switch layout for events from grid to list view (#1044)
## Description

This PR adds the ability to switch from grid view to list view. Finding
an event in grid view can be difficult because it's hard for the eye to
scan the events.

Setting an option is cached. I also added analytics for this button.

## Demo


https://github.com/SharezoneApp/sharezone-app/assets/24459435/4742669a-1f0b-4d1e-a20c-af176b324fb6

(Don't be confused that there are events from the past. For the demo I
changed the query to load all events to have more events on my page).