Fix showing error message when homework `todoUntil` field is null (#1053)
## Description

Normally, the todo until field is calculated when the user selects a
course. However, there is the edge case where this not work and the todo
until field is still null after selecting a course. In this case, a null
value exception was thrown (see #1052). This PR fixes this, throws a
helpful exception and makes the "Datum auswählen" text red.

## Demo


https://github.com/SharezoneApp/sharezone-app/assets/24459435/5d2c67a8-d98f-4d23-adc2-589d9b7bbc50

## Related Tickets

Fixes #1052